done hitsound exporter class

+++ Mapping_Tools_Core/BeatmapHelper/Editor/BeatmapEditor.cs
@@ -11,9 +11,10 @@ namespace Mapping_Tools_Core.BeatmapHelper.Editor {
         /// Saves the beatmap using <see cref=".SaveFile()"/> but also updates the filename according to the metadata of the <see cref="Beatmap"/>
         /// </summary>
         /// <remarks>This method also updates the Path property</remarks>
-        public void WriteFileWithNameUpdate(Beatmap beatmap) {
+        public void WriteFileWithNameUpdate(Beatmap beatmap, bool deleteOriginal = true) {
             // Remove the beatmap with the old filename
-            File.Delete(Path);
+            if (deleteOriginal)
+                File.Delete(Path);
 
             // Save beatmap with the new filename
             Path = System.IO.Path.Combine(GetParentFolder(), beatmap.GetFileName());
